Modify icons for musescore files
Application Name: MuseScore
Icon Name: application-x-musescore.svg

This icon uses the brand color for MuseScore https://musescore.org/
I use MuseScore program on daily basis and I don't like that the papirus icon for this mimetype is white. I think it should be with the default blue color and the app icon.

This is my first contribution so I hope i did it all right.