remove ePapirus form install.sh and readme
There is no need for ePapirus now, it is in fact now combined with Papirus/Papirus-Dark. It was originally created because of the different sizes of "-symbolic" icons in Gnome and Pantheon and for tray too. Tray removed from pantheon and there's no point in maintaining it in the future. Now this problem has been solved with the unification of the "-symbolic" icons. I hope so) XFCE users can switch to original Papirus. Need tests, different DE's