Delete ePapirus and ePapirus-Dark icon themes
ePapirus is no longer needed as it will be merged into Papirus. It was
originally created due to differences in "-symbolic" icon sizes between
GNOME and Pantheon, as well as for tray support. Since Pantheon has
removed the tray and the symbolic icon issue has been resolved,
maintaining ePapirus is unnecessary.